Work around broken HAVE_BYTES detection
configure uses ocamlc -safe-string to determine whether the compiler has
the bytes type. However, on OCaml 5, this command, without input files,
exits with a non-zero exit code and prints "No input files". This causes
configure to wrongly believe that -safe-string is not supported, which
then triggers a multitude of compilation errors.

This commit changes the code to unconditionally assume that bytes is
available. A better way is probably to adapt OASIS, or what generated
the configure script.